Rewrited index.ts functions
