+++ package.json
@@ -17,7 +17,7 @@
   "devDependencies": {
     "@types/node": "^8.5.8",
     "@types/uws": "^0.13.1",
-    "ts-builder": "^0.6.0"
+    "ts-builder": "^0.6.4"
   },
   "dependencies": {
     "uws": "^9.14.0"
