+++ src/utils/functions.ts
@@ -1,17 +1,20 @@
-import * as crypto from 'crypto'
+import { randomBytes } from 'crypto'
+import { CustomObject } from './types'
 
 export function logError<T>(data: T): any {
-    return console.log('\x1b[31m%s\x1b[0m', data)
+  return console.log(`\x1b[31m${data}\x1b[0m`)
 }
 
-export function logReady<T>(data: T): any {
-    return console.log('\x1b[36m%s\x1b[0m', data)
+export function logReady<T>(data: T, stamp: boolean = false): any {
+  return console.log(`\x1b[36m${data}\x1b[0m`)
 }
 
 export function logWarning<T>(data: T): any {
-    return console.log('\x1b[33m%s\x1b[0m', data)
+  return console.log(`\x1b[33m${data}\x1b[0m`)
 }
 
 export function generateKey(length: number): string {
-    return crypto.randomBytes(Math.ceil(length / 2)).toString('hex').slice(0, length)
+  return randomBytes(Math.ceil(length / 2))
+    .toString('hex')
+    .slice(0, length)
 }
