+++ src1/modules/worker/socket/parser.ts
@@ -6,7 +6,7 @@ export function encode(event: string, data: any, type: string): string {
         'emit': { '#': ['e', event, data] },
         'publish': { '#': ['p', event, data] },
         'system': {
-            'subsribe': { '#': ['s', 's', data] },
+            'subscribe': { '#': ['s', 's', data] },
             'unsubscribe': { '#': ['s', 'u', data] },
             'configuration': { '#': ['s', 'c', data] }
         }
